Tested and fixed all user management features

+++ migrations/20160715110829-create-user.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       phone: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT(11)
       },
       email: {
         allowNull: false,
